+++ chaosazure/machine/actions.py
@@ -1,25 +1,28 @@
 # -*- coding: utf-8 -*-
 import random
+import secrets
+from typing import List, Any
 
+from chaoslib import configuration
 from chaoslib.types import Configuration, Secrets
+from knack import config
 from logzero import logger
 
 from chaosazure.cli.runner import execute
 from chaosazure.machine.commands import restart_machine_command, \
-    poweroff_machine_command, delete_machine_command
+    poweroff_machine_command, delete_machine_command, start_machine_command
 from chaosazure.machine.picker import pick_machines
 
-__all__ = ["delete_machine", "poweroff_machine", "restart_machine"]
+__all__ = ["delete_machine", "poweroff_machine", "restart_machine",
+           "start_machine"]
+
+stopped_machines: List[Any] = []
 
 
 def delete_machine(configuration: Configuration = None,
                    secrets: Secrets = None, filter: str = None):
     """
     Delete a virtual machines at random.
-
-    ***Be aware**: Deleting a machine is an invasive action. You will not be
-    able to recover the machine once you deleted it.
-
     Parameters
     ----------
     filter : str
@@ -42,7 +45,6 @@ def poweroff_machine(configuration: Configuration = None,
                      secrets: Secrets = None, filter: str = None):
     """
     Power off a virtual machines at random.
-
     Parameters
     ----------
     filter : str
@@ -57,6 +59,8 @@ def poweroff_machine(configuration: Configuration = None,
     machines = pick_machines(configuration, secrets, filter)
     choice = random.choice(machines)
 
+    stopped_machines.append(choice)
+
     command = poweroff_machine_command(choice)
     execute(configuration, secrets, command)
 
@@ -65,7 +69,6 @@ def restart_machine(configuration: Configuration = None,
                     secrets: Secrets = None, filter: str = None):
     """
     Restart a virtual machines at random.
-
     Parameters
     ----------
     filter : str
@@ -82,3 +85,20 @@ def restart_machine(configuration: Configuration = None,
 
     command = restart_machine_command(choice)
     execute(configuration, secrets, command)
+
+
+def start_machine(configuration: Configuration = None,
+                  secrets: Secrets = None):
+    """
+        Start all stopped virtual machines.
+    """
+
+    logger.debug("Start start_machine: configuration='{}'"
+                 .format(configuration))
+    logger.debug(
+        "Found machines: {}".format([x['name'] for x in stopped_machines]))
+    for machine in stopped_machines:
+        command = start_machine_command(machine)
+        execute(configuration, secrets, command)
+
+    stopped_machines.clear()