+++ tests/machine/test_machine_commands.py
@@ -1,5 +1,5 @@
 from chaosazure.machine.commands import delete_machine_command, \
-    poweroff_machine_command, restart_machine_command
+    poweroff_machine_command, restart_machine_command, start_machine_command
 
 TESTED_RESOURCE = "Microsoft.ContainerService/managedClusters"
 
@@ -39,5 +39,17 @@ def test_restart_machine_command():
 
     assert len(command) == 6
     assert command[1] == 'restart'
+    assert command[3] == machine['resourceGroup']
+    assert command[5] == machine['name']
+
+
+def test_start_machine_command():
+    machine = {
+        'name': 'myname',
+        'resourceGroup': 'myresourcegroup'
+    }
+    command = start_machine_command(machine)
+    assert len(command) == 6
+    assert command[1] == 'start'
     assert command[3] == machine['resourceGroup']
     assert command[5] == machine['name']