+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 -   Introduce actions on Azure Virtual Machine Scalesets (VMSS), like stopping, restarting, and
     deallocating AKS nodes
 -   Emphasize the risk of deleting actions in the docstrings
+-   Introduce rollback action for virtual machines
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.2.0...HEAD